UserId error solving
